Rename piper_control_node to piper_control_ros2 to avoid circular import issue when using piper_control_node file
