+++ pixi.toml
@@ -18,5 +18,5 @@ ros-jazzy-desktop = ">=0.11.0,<0.12"
 [pypi-dependencies]
 piper_control           = { git = "https://github.com/Reimagine-Robotics/piper_control/", branch = "main" }
 piper_ros               = { path = ".", editable = true }
-piper_control_node      = { path = "src/piper_control_node", editable = true }
+piper_control_ros2      = { path = "src/piper_control_ros2", editable = true }
 piper_puppeteering_node = { path = "src/piper_puppeteering_node", editable = true }
