--- src/piper_control_node/piper_control_node/__init__.py
@@ -1,3 +1,0 @@
-"""piper_ros.piper_control_node package."""
-
-from ._version import __version__
