--- src/piper_control_node/piper_control_node/_version.py
@@ -1,6 +1,0 @@
-"""Version information for the piper_control_node package.
-
-Update this file manually when there is a change to the piper_control_node.
-"""
-
-__version__ = "1.0.2"
