--- src/piper_control_node/setup.cfg
@@ -1,4 +1,0 @@
-# [develop]
-# script_dir=$base/lib/piper_control_node
-[install]
-install_scripts=$base/lib/piper_control_node
