+++ src/piper_control_ros2/piper_control_ros2/__init__.py
@@ -1,0 +1,3 @@
+"""piper_ros.piper_control_ros2 package."""
+
+from ._version import __version__
