+++ src/piper_control_ros2/piper_control_ros2/_version.py
@@ -1,0 +1,6 @@
+"""Version information for the piper_control_ros2 package.
+
+Update this file manually when there is a change to the piper_control_ros2.
+"""
+
+__version__ = "1.0.2"
