+++ src/piper_control_ros2/setup.cfg
@@ -1,0 +1,4 @@
+# [develop]
+# script_dir=$base/lib/piper_control_ros2
+[install]
+install_scripts=$base/lib/piper_control_ros2
