+++ src/piper_control_ros2/setup.py
@@ -1,4 +1,4 @@
-# Setup script for the piper_control_node package.
+# Setup script for the piper_control_ros2 package.
 #
 # Note: We install an executable wrapper script manually to work around ROS2’s
 # default behavior of hardcoding `#!/usr/bin/python3` as the shebang in
@@ -9,13 +9,13 @@
 # python` as its shebang, we ensure that ROS2 launches the node using the
 # currently active environment's Python interpreter (e.g., from Conda).
 #
-# This allows us to `ros2 run piper_control_node run_piper_control_node` while
+# This allows us to `ros2 run piper_control_ros2 run_piper_control_node` while
 # correctly using the Conda environment for dependencies.
 
 import os
 from setuptools import find_packages, setup
 
-package_name = "piper_control_node"
+package_name = "piper_control_ros2"
 
 setup(
     name=package_name,
@@ -29,7 +29,7 @@ setup(
         ("share/" + package_name, ["package.xml"]),
         (
             os.path.join("share", package_name, "data"),
-            ["piper_control_node/teach_mode/piper_grav_comp.xml"],
+            ["piper_control_ros2/teach_mode/piper_grav_comp.xml"],
         ),
     ],
     install_requires=["setuptools"],
@@ -40,7 +40,7 @@ setup(
     license="MIT",
     entry_points={
         "console_scripts": [
-            "piper_control_node = piper_control_node.piper_control_node:main",
+            "piper_control_ros2 = piper_control_ros2.piper_control_node:main",
         ],
     },
 )
